+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Changed
 - [MINOR] Updated `Requester` to send data correctly for PUT requests
+- [MINOR] Updated `Requester` to send data correctly for PATCH requests
 
 ### Removed